pyproject.toml
[tool.poetry]
name = "homeassistant-vacuum-viomi"
version = "0.0.4"
description = "Integration of a Xiaomi Viomi vacuum devices"
authors = ["Artem Poliukhovych <me@2mio.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-miio = "^0.5.9"
voluptuous = "^0.12.1"
construct = "^2.10.67"
homeassistant = "^2021.12.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
mypy = "^0.910"
pytest-cov = "^2.12.1"
black = "^21.6b0"
isort = "^5.9.1"
pytest-homeassistant-custom-component = "^0.5.12"
micloud = "^0.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
